Lift leader group hints up
Rather than having each group have its own hint, which the parent can retrieve,
the parent has a mapping of children => hints. This makes it easier to have
hints on a group which e.g. has a command in it, because it means that the group
can have a hint for the command even if the command itself doesn't have a hint.