Provide interface for leaders and UI to talk
The brief idea is that the UI provides a couple of basic ways for leaders to
configure it (primarily adding and removing hints, and setting the type of input
the UI should expect) while the UI is expected to coordinate the among the
choice of leaders. LeaderUI codifies the boundaries of this interaction.

When implementing this, LeaderAction grew the ability to launch its command. For
testing purposes, it can't launch its command on its own, but instead delegates
to a handler which accepts the command and the arguments. Calls to the
constructors that create actions had to be modified to provide the runner.